refactor(authentication): improve abstraction of oauth2 settings refactor(authentication): abstract oauth2 provider initialization
