--- modules/authentication/src/handlers/AuthenticationProviders/interfaces/OAuth2Settings.ts
@@ -1,16 +1,0 @@
-export interface OAuth2Settings {
-  providerName: string;
-  authorizeUrl: string;
-  tokenUrl: string;
-  callbackUrl: string;
-  finalRedirect: string;
-  accountLinking: boolean;
-  clientId: string;
-  clientSecret: string;
-  accessTokenMethod: string;
-  grantType?: string;
-  state?: string;
-  scope?: string;
-  responseType?: string;
-  scopeSeperator?: string;
-}
