--- modules/authentication/src/handlers/facebook/facebook.user.ts
@@ -1,5 +1,0 @@
-import { Payload } from '../AuthenticationProviders/interfaces/Payload';
-
-export interface FacebookUser extends Payload {
-  [key:string]: any;
-}
