--- modules/authentication/src/handlers/figma/figma.user.ts
@@ -1,8 +1,0 @@
-import { Payload } from '../../handlers/AuthenticationProviders/interfaces/Payload';
-
-export interface FigmaUser extends Payload {
-  data: {
-    handle?: string;
-    image_url?: string;
-  }
-}
