--- modules/authentication/src/handlers/google/google.user.ts
@@ -1,12 +1,0 @@
-import { Payload } from '../AuthenticationProviders/interfaces/Payload';
-
-export interface GoogleUser extends Payload {
-  data: {
-    name?: string;
-    given_name?: string;
-    family_name?: string;
-    picture?: string;
-    locale?: string;
-    verified_email?: boolean
-  },
-}
