+++ modules/authentication/src/handlers/oauth2/facebook/facebook.user.ts
@@ -1,0 +1,5 @@
+import { Payload } from '../interfaces/Payload';
+
+export interface FacebookUser extends Payload {
+  [key:string]: any;
+}
