+++ modules/authentication/src/handlers/oauth2/figma/figma.json
@@ -1,0 +1,7 @@
+{
+  "accessTokenMethod": "POST",
+  "authorizeUrl": "https://www.figma.com/oauth?",
+  "tokenUrl": "https://www.figma.com/api/oauth/token",
+  "responseType": "code",
+  "grantType": "authorization_code"
+}
