+++ modules/authentication/src/handlers/oauth2/figma/figma.user.ts
@@ -1,0 +1,8 @@
+import { Payload } from '../interfaces/Payload';
+
+export interface FigmaUser extends Payload {
+  data: {
+    handle?: string;
+    image_url?: string;
+  }
+}
