+++ modules/authentication/src/handlers/oauth2/github/github.json
@@ -1,0 +1,7 @@
+{
+  "accessTokenMethod": "POST",
+  "authorizeUrl": "https://github.com/login/oauth/authorize?",
+  "providerName": "github",
+  "tokenUrl": "https://github.com/login/oauth/access_token",
+  "responseType": "code"
+}
