+++ modules/authentication/src/handlers/oauth2/google/google.json
@@ -1,0 +1,8 @@
+{
+  "accessTokenMethod": "POST",
+  "providerName": "google",
+  "tokenUrl": "https://oauth2.googleapis.com/token",
+  "responseType": "code",
+  "grantType": "authorization_code",
+  "authorizeUrl": "https://accounts.google.com/o/oauth2/v2/auth?"
+}
