+++ modules/authentication/src/handlers/oauth2/index.ts
@@ -1,0 +1,7 @@
+export * from "./facebook/facebook";
+export * from "./figma/figma";
+export * from "./github/github";
+export * from "./google/google";
+export * from "./microsoft/microsoft";
+export * from "./slack/slack";
+export * from "./twitch/twitch";
