+++ modules/authentication/src/handlers/oauth2/interfaces/AuthParams.ts
