+++ modules/authentication/src/handlers/oauth2/interfaces/RedirectOptions.ts
