+++ modules/authentication/src/handlers/oauth2/microsoft/microsoft.json
@@ -1,0 +1,9 @@
+{
+  "accessTokenMethod": "POST",
+  "authorizeUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/authorize?",
+  "providerName": "microsoft",
+  "tokenUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/token",
+  "response_mode": "form_post",
+  "responseType": "code",
+  "grantType": "authorization_code"
+}
