+++ modules/authentication/src/handlers/oauth2/microsoft/microsoft.user.ts
@@ -1,0 +1,15 @@
+import { Payload } from '../interfaces/Payload';
+
+export interface MicrosoftUser extends Payload {
+  data: {
+    displayName?: string;
+    givenName?: string;
+    jobTitle?: string;
+    surname?: string;
+    userPrincipalName?: string;
+    mobilePhone?: string;
+    officeLocation?: string;
+    preferredLanguage?: string;
+    businessPhones?: string[];
+  }
+}
