+++ modules/authentication/src/handlers/oauth2/slack/slack.json
@@ -1,0 +1,7 @@
+{
+  "accessTokenMethod" : "POST",
+  "authorizeUrl" : "https://slack.com/oauth/authorize?",
+  "providerName" : "slack",
+  "tokenUrl" : "https://slack.com/api/oauth.access",
+  "responseType" : "code"
+}
