+++ modules/authentication/src/handlers/oauth2/twitch/twitch.json
@@ -1,0 +1,8 @@
+{
+  "accessTokenMethod" : "POST",
+  "authorizeUrl" : "https://id.twitch.tv/oauth2/authorize?",
+  "providerName" : "twitch",
+  "tokenUrl" : "https://id.twitch.tv/oauth2/token",
+  "grantType" : "authorization_code",
+  "responseType" : "code"
+}
