+++ modules/authentication/src/handlers/oauth2/twitch/twitch.user.ts
@@ -1,0 +1,14 @@
+import { Payload } from '../interfaces/Payload';
+
+export interface TwitchUser extends Payload {
+  data: {
+    display_name?: string;
+    profile_image_url?: string;
+    offline_image_url?: string;
+    login?: string;
+    broadcaster_type?: string;
+    type?: string;
+    description?: string;
+    view_count?: string;
+  }
+}
